dom/tests/mochitest/general/test_bug1161721.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1161721
-->
<head>
  <title>Test for Bug 1161721</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/WindowSnapshot.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1161721">Mozilla Bug 1161721</a>
<p id="display"></p>

<div id="content">
</div>

<pre id="test">
  <script type="application/javascript">
    ok(!document.queryCommandSupported("paste"), "Paste isn't supported in non-privilged JavaScript");
    ok(document.queryCommandSupported("copy"), "Copy is supported in non-privilged JavaScript");
    ok(document.queryCommandSupported("cut"), "Cut is supported in non-privilged JavaScript");

    ok(SpecialPowers.wrap(document).queryCommandSupported("paste"), "Paste is supported in privilged JavaScript");
    ok(SpecialPowers.wrap(document).queryCommandSupported("copy"), "Copy is supported in privilged JavaScript");
    ok(SpecialPowers.wrap(document).queryCommandSupported("cut"), "Cut is supported in privilged JavaScript");
  </script>
</pre>
</body>
</html>
